JWT 토큰 만료 예외처리 필터 추가 / 전역 예외처리 예외 타입 수정 (ApiRequestException 추가) #32 #40

+++ src/main/java/com/udangtangtang/backend/service/ArticleService.java
@@ -2,6 +2,7 @@ package com.udangtangtang.backend.service;
 
 import com.udangtangtang.backend.domain.*;
 import com.udangtangtang.backend.dto.LocationRequestDto;
+import com.udangtangtang.backend.exception.ApiRequestException;
 import com.udangtangtang.backend.repository.*;
 import com.udangtangtang.backend.util.LocationDataPreprocess;
 import lombok.RequiredArgsConstructor;
@@ -42,7 +43,7 @@ public class ArticleService {
 
     public Article getArticle(Long id) {
         return articleRepository.findById(id).orElseThrow(
-                () -> new NullPointerException(String.format("해당되는 아이디(%d)의 게시물이 없습니다.", id))
+                () -> new ApiRequestException(String.format("해당되는 아이디(%d)의 게시물이 없습니다.", id))
         );
     }
 
@@ -66,19 +67,21 @@ public class ArticleService {
     }
 
     @Transactional
-    public Article updateArticle(User user, Long id, String text, LocationRequestDto locationRequestDto, List<String> tagNames, List<MultipartFile> imageFiles, List<Long> rmImageIdList) {
+    public Article updateArticle(User user, Long id, String text, LocationRequestDto locationRequestDto, List<String> tagNames, List<MultipartFile> imageFiles, List<Long> rmImageIds) {
         Article article = articleRepository.findById(id).orElseThrow(
-                () -> new NullPointerException(String.format("해당되는 아이디(%d)의 게시물이 없습니다.", id))
+                () -> new ApiRequestException(String.format("해당되는 아이디(%d)의 게시물이 없습니다.", id))
         );
 
         // 기존에 저장된 이미지 삭제
-        for (Long imageId : rmImageIdList) {
-            Image image = imageRepository.findById(imageId).orElseThrow(
-                    () -> new NullPointerException(String.format("해당되는 아이디(%d)의 이미지가 없습니다.", imageId))
-            );
-            fileProcessService.deleteImage(image.getUrl());
+        if(rmImageIds != null) {
+            for (Long imageId : rmImageIds) {
+                Image image = imageRepository.findById(imageId).orElseThrow(
+                        () -> new ApiRequestException(String.format("해당되는 아이디(%d)의 이미지가 없습니다.", imageId))
+                );
+                fileProcessService.deleteImage(image.getUrl());
+            }
+            imageRepository.deleteAllById(rmImageIds);
         }
-        imageRepository.deleteAllById(rmImageIdList);
 
         locationDataPreprocess.categoryNamePreprocess(locationRequestDto);
         Location location = locationRepository.save(new Location(locationRequestDto, user.getId()));
@@ -101,9 +104,9 @@ public class ArticleService {
     }
 
     @Transactional
-    public void deleteArticle(Long id) throws IllegalArgumentException {
+    public void deleteArticle(Long id) {
         Article article = articleRepository.findById(id).orElseThrow(
-                () -> new NullPointerException(String.format("아이디(%d)에 해당되는 게시물이 없습니다.", id))
+                () -> new ApiRequestException(String.format("아이디(%d)에 해당되는 게시물이 없습니다.", id))
         );
 
         // S3에 업로드된 이미지 삭제
